requirements.txt
streamlit
pandas
requests
beautifulsoup4
nltk
scikit-learn
numpy
matplotlib
seaborn
wordcloud
plotly
fpdf
bertopic
umap-learn
hdbscan
